+++ LOVE-WrapLua/OneLua/graphics_psp.lua
@@ -13,6 +13,7 @@ end
 
 lv1lua.loadOnce("LOVE-WrapLua/core/util.lua")
 lv1lua.loadOnce("LOVE-WrapLua/core/textwrap.lua")
+lv1lua.loadOnce("LOVE-WrapLua/core/polyfill.lua")
 
 local GRAPHICS = "LOVE-WrapLua/OneLua/psp/"